requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
torch==2.1.2
transformers==4.36.2
langchain==0.1.1
langchain-google-genai==0.0.5
langchain-qdrant==0.0.2
langchain-community==0.0.13
python-dotenv==1.0.0
qdrant-client==1.7.0
jinja2==3.1.2